Refactor code generation request handling
