+++ packages/server/api/src/app/ai/chat/ai-mcp-chat.controller.ts
@@ -2,7 +2,6 @@ import { FastifyPluginAsyncTypebox } from '@fastify/type-provider-typebox';
 import { logger } from '@openops/server-shared';
 import {
   ApplicationError,
-  ChatFlowContext,
   ChatNameRequest,
   CODE_BLOCK_NAME,
   DeleteChatHistoryRequest,
@@ -16,7 +15,6 @@ import {
 import { CoreMessage } from 'ai';
 import { FastifyReply } from 'fastify';
 import { StatusCodes } from 'http-status-codes';
-import { v4 as uuidv4 } from 'uuid';
 import {
   createChatContext,
   deleteChatHistory,
@@ -32,6 +30,7 @@ import {
   saveChatHistory,
   updateChatName,
 } from './ai-chat.service';
+import { handleCodeGenerationRequest } from './code-generation-handler';
 import { streamCode } from './code.service';
 import { enrichContext, IncludeOptions } from './context-enrichment.service';
 import { getBlockSystemPrompt } from './prompts.service';
@@ -140,32 +139,31 @@ export const aiMCPChatController: FastifyPluginAsyncTypebox = async (app) => {
       const isCodeGenerationRequest =
         chatContext?.blockName === CODE_BLOCK_NAME;
 
+      const newMessage: CoreMessage = {
+        role: 'user',
+        content: messageContent,
+      };
+
+      const generationRequestParams = {
+        app,
+        authToken,
+        chatId,
+        userId,
+        projectId,
+        newMessage,
+        additionalContext: request.body.additionalContext,
+        serverResponse: reply.raw,
+      };
+
       if (isCodeGenerationRequest) {
         logger.debug('Using code generation flow');
-        const response = await handleCodeGenerationRequest({
-          chatId,
-          userId,
-          projectId,
-          messageContent,
-          additionalContext: request.body.additionalContext,
-        });
+        const response = await handleCodeGenerationRequest(
+          generationRequestParams,
+        );
         return response;
       } else {
         logger.debug('Using normal conversation flow');
-        const newMessage: CoreMessage = {
-          role: 'user',
-          content: messageContent,
-        };
-
-        await handleUserMessage({
-          app,
-          chatId,
-          userId,
-          projectId,
-          authToken,
-          newMessage,
-          serverResponse: reply.raw,
-        });
+        await handleUserMessage(generationRequestParams);
         return;
       }
     } catch (error) {
@@ -368,226 +366,6 @@ function handleError(
 
   logger.error(`Failed to process ${context || 'request'} with error: `, error);
   return reply.code(500).send({ message: 'Internal server error' });
-}
-
-/**
- * Handles code generation requests using streamCode for structured output.
- */
-async function handleCodeGenerationRequest({
-  chatId,
-  userId,
-  projectId,
-  messageContent,
-  additionalContext,
-}: {
-  chatId: string;
-  userId: string;
-  projectId: string;
-  messageContent: string;
-  additionalContext?: ChatFlowContext;
-}): Promise<Response> {
-  try {
-    const conversationResult = await getConversation(chatId, userId, projectId);
-    const llmConfigResult = await getLLMConfig(projectId);
-
-    conversationResult.messages.push({
-      role: 'user',
-      content: messageContent,
-    });
-
-    const { chatContext, messages } = conversationResult;
-    const { aiConfig, languageModel } = llmConfigResult;
-
-    const enrichedContext = additionalContext
-      ? await enrichContext(additionalContext, projectId, {
-          includeCurrentStepOutput: IncludeOptions.ALWAYS,
-        })
-      : undefined;
-
-    const prompt = await getBlockSystemPrompt(chatContext, enrichedContext);
-
-    logger.debug('Code generation prompt:', {
-      chatContext: JSON.stringify(chatContext),
-      enrichedContext: enrichedContext ? 'present' : 'absent',
-      promptLength: prompt.length,
-      promptPreview: prompt.substring(0, 200),
-    });
-
-    let codeResult: {
-      type: string;
-      code?: string;
-      packageJson?: string;
-      textAnswer: string;
-    } | null = null;
-
-    const result = streamCode({
-      messages,
-      languageModel,
-      aiConfig,
-      systemPrompt: prompt,
-      onFinish: async (result) => {
-        if (result.object) {
-          codeResult = result.object;
-        }
-      },
-      onError: (error) => {
-        logger.error('Failed to generate code', {
-          error,
-        });
-        throw error;
-      },
-    });
-
-    // Wait for the stream to complete and get the result
-    const streamResponse = result.toTextStreamResponse();
-    const reader = streamResponse.body?.getReader();
-    if (reader) {
-      let done = false;
-      while (!done) {
-        const result = await reader.read();
-        done = result.done;
-      }
-    }
-
-    // Wait a bit for onFinish to be called
-    await new Promise((resolve) => setTimeout(resolve, 100));
-
-    if (!codeResult) {
-      throw new Error('No code generation result received');
-    }
-
-    // Create a mock tool call response
-    const toolCallId = uuidv4();
-
-    // Create the streaming response manually
-    const encoder = new TextEncoder();
-    const stream = new ReadableStream({
-      start(controller) {
-        // Send tool call start (use 9: prefix for tool calls)
-        const toolCallMessage = {
-          toolCallId,
-          toolName: 'generate_code',
-          args: { message: messageContent },
-        };
-        controller.enqueue(
-          encoder.encode(`9:${JSON.stringify(toolCallMessage)}\n`),
-        );
-
-        // Send tool call result
-        const toolResult = {
-          toolCallId,
-          result: {
-            code:
-              (codeResult as { code?: string; packageJson?: string }).code ||
-              '',
-            packageJson:
-              (codeResult as { code?: string; packageJson?: string })
-                .packageJson || '{}',
-          },
-        };
-        controller.enqueue(encoder.encode(`a:${JSON.stringify(toolResult)}\n`));
-
-        // Send finish message
-        const finishMessage = {
-          finishReason: 'tool-calls',
-          usage: { promptTokens: null, completionTokens: null },
-          isContinued: false,
-        };
-        controller.enqueue(
-          encoder.encode(`e:${JSON.stringify(finishMessage)}\n`),
-        );
-        controller.enqueue(
-          encoder.encode(`d:${JSON.stringify(finishMessage)}\n`),
-        );
-
-        // Send text content as part of the assistant message
-        controller.enqueue(
-          encoder.encode(
-            `0:"${(codeResult as { textAnswer: string }).textAnswer}"\n`,
-          ),
-        );
-
-        // Send finish for text message
-        const textFinishMessage = {
-          finishReason: 'stop',
-          usage: { promptTokens: null, completionTokens: null },
-          isContinued: false,
-        };
-        controller.enqueue(
-          encoder.encode(`e:${JSON.stringify(textFinishMessage)}\n`),
-        );
-        controller.enqueue(
-          encoder.encode(`d:${JSON.stringify(textFinishMessage)}\n`),
-        );
-
-        controller.close();
-      },
-    });
-
-    // Save the conversation history in the correct format for mergeToolResultsIntoMessages
-    const saveToolCallId = uuidv4();
-
-    // Save assistant message with tool call and text (result will be merged by mergeToolResultsIntoMessages)
-    const assistantMessage: CoreMessage = {
-      role: 'assistant',
-      content: [
-        {
-          type: 'tool-call',
-          toolCallId: saveToolCallId,
-          toolName: 'generate_code',
-          args: { message: messageContent },
-        },
-        {
-          type: 'text',
-          text: (codeResult as { textAnswer: string }).textAnswer || '',
-        },
-      ],
-    };
-
-    // Save tool result message (will be merged into assistant message by mergeToolResultsIntoMessages)
-    const toolResultMessage: CoreMessage = {
-      role: 'tool',
-      content: [
-        {
-          type: 'tool-result',
-          toolCallId: saveToolCallId,
-          toolName: 'generate_code',
-          result: {
-            content: [
-              {
-                type: 'text',
-                text: JSON.stringify({
-                  code:
-                    (codeResult as { code?: string; packageJson?: string })
-                      .code || '',
-                  packageJson:
-                    (codeResult as { code?: string; packageJson?: string })
-                      .packageJson || '{}',
-                }),
-              },
-            ],
-            isError: false,
-          },
-        },
-      ],
-    };
-
-    await saveChatHistory(chatId, userId, projectId, [
-      ...messages,
-      assistantMessage,
-      toolResultMessage,
-    ]);
-
-    return new Response(stream, {
-      headers: {
-        'Content-Type': 'text/plain; charset=utf-8',
-        'Transfer-Encoding': 'chunked',
-      },
-    });
-  } catch (error) {
-    logger.error('Error in handleCodeGenerationRequest:', error);
-    throw error;
-  }
 }
 
 /**
